Extracted Patch and Replace to their own files
