+++ enterprise/internal/scim/user_replace.go
@@ -1,0 +1,20 @@
+package scim
+
+import (
+	"net/http"
+
+	"github.com/elimity-com/scim"
+	"github.com/sourcegraph/log"
+)
+
+// Replace replaces ALL existing attributes of the resource with given identifier. Given attributes that are empty
+// are to be deleted. Returns a resource with the attributes that are stored.
+func (h *UserResourceHandler) Replace(r *http.Request, id string, attributes scim.ResourceAttributes) (scim.Resource, error) {
+	// TODO: Add real logic
+	attributesString := resourceAttributesToLoggableString(attributes)
+	h.observationCtx.Logger.Error("XXXXX Replace", log.String("method", r.Method), log.String("id", id), log.String("attributes", attributesString))
+
+	return scim.Resource{
+		ID: "123",
+	}, nil
+}
